Updated project to 4.0, added several moons, and added NGUI
Not all of the moons are spherical, so I added those that were.  We'll
need to ask about the rest, but I added the ones that had maps
available.  I've added NGUI as well, although I haven't done anything
with it.  Any NGUI objects that are doing something need an
NGUIEventPasser attached.  I have not created a window event handler at
this time, so the NGUI passer won't have anywhere to send to yet.